fix(secrets): the declared key mode is not applicable on Windows, not unchecked
test (windows-latest) failed TestFileAuthority_PresentAndCorrectModeIsOK and
TestFileAuthority_WrongModeFails, and the tests were right to fail: the check
was asking Windows for a property it does not keep.

Windows has no Unix permission bits. os.Chmod toggles the read-only attribute
and nothing else; fi.Mode().Perm() reports a synthesised 0666 or 0444 that no
caller set. Comparing that to 0600 would report FAILED on every healthy Windows
machine — a gate wrong about a whole class of machine teaches its readers to
discount it, which is GUARD-002's failure arrived at from the other side.

So checkKeyMode is build-tagged, and the Windows half returns nil with the
reason stated in its own file rather than inferred from the Unix one. This is
'not applicable', not 'not checked': there is no Windows property the
declaration describes. The checks that DO carry meaning on both platforms —
present, regular file, non-empty — still run there, asserted by a Windows-only
test so the claim is behaviour and not a comment.

If the ACL equivalent is ever wanted, it is a real check with a real subject and
belongs in that file as one, not as a reinterpretation of a number Windows does
not keep.
