+++ cli/internal/secrets/fileauthority_unix_test.go
@@ -3,6 +3,8 @@
 package secrets
 
 import (
+	"errors"
+	"os"
 	"path/filepath"
 	"strings"
 	"syscall"
@@ -31,5 +33,40 @@ func TestFileAuthority_NonRegularFileFails(t *testing.T) {
 	}
 	if !strings.Contains(err.Error(), "regular file") {
 		t.Errorf("the error must name the actual problem, got: %v", err)
+	}
+}
+
+// The two mode cases live here for the same reason as the FIFO one: Windows has no
+// Unix permission bits, so a 0644 key cannot exist there to be rejected. The
+// platform's own behaviour is asserted in fileauthority_windows_test.go instead.
+
+func TestFileAuthority_WrongModeFails(t *testing.T) {
+	p := writeKey(t, 0o644)
+	// Confirm the mutation landed: a umask could have produced 0600 and the
+	// assertion below would then pass for the wrong reason.
+	fi, err := os.Stat(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if got := fi.Mode().Perm(); got != 0o644 {
+		t.Fatalf("fixture is mode %04o, not the 0644 this case is about", got)
+	}
+
+	err = (&Loader{}).Verify(rootEntry(t, p, 0o600))
+	if err == nil {
+		t.Fatal("a world-readable root must FAIL, not pass")
+	}
+	if errors.Is(err, ErrSecretAbsent) {
+		t.Fatal("a wrong mode is a defect, not an absence — reporting MISSING hides it")
+	}
+	if !strings.Contains(err.Error(), "0644") || !strings.Contains(err.Error(), "0600") {
+		t.Errorf("the error must name what it found and what it wanted, got: %v", err)
+	}
+}
+
+func TestFileAuthority_DefaultsToO600WhenNoModeDeclared(t *testing.T) {
+	p := writeKey(t, 0o644)
+	if err := (&Loader{}).Verify(rootEntry(t, p, 0)); err == nil {
+		t.Fatal("mode 0 means 'the 0600 default', so 0644 must still FAIL")
 	}
 }
